Change string to single character to try and get game working
I had double quotation instead of single quotation and this was turning the zero into a string.  I had to remove the double quotes and replace w single quotes so that I can make the string into a char so the comparison will be between same data types.  a comparison wont work here for different data types comparing a char to a string this gave an error